removed trailing whitespaces and use more structured bindings

+++ string/lyndon.cpp
@@ -1,5 +1,5 @@
 bool next(string& s, int n, char mi = '0', char ma = '1') {
-	for (ll i = sz(s), j = sz(s); i < n; i++)
+	for (int i = sz(s), j = sz(s); i < n; i++)
 		s.push_back(s[i % j]);
 	while(!s.empty() && s.back() == ma) s.pop_back();
 	if (s.empty()) {
